feat!: completely re-implement the HttpClient (abstraction) using Undici
This is heavily based on the new HttpClient code I wrote
for https://github.com/SpraxDev/Api.Sprax2013.De
The abstraction here might need to change a bit to allow
for the special case of proxy(-pool) http clients that
SpraxAPI uses. I am not sure if this implementation allows
for such deep control.

I tried having an abstraction this time so the actual http client
*could* be changed as desired.

This also gets rid of `superagent` as dependency, which
was deprecated/abandoned for a while now.